requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-chroma==0.1.3
langchain-groq==0.1.9
langchain-hub==0.1.15
langgraph==0.2.24
pypdf==4.3.1
huggingface_hub==0.24.6
chromadb==0.5.5
python-dotenv==1.0.1
streamlit==1.38.0
langchain-huggingface==0.0.3
torch==2.4.0
sentence-transformers==3.0.1